% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega.R
\name{simple_wordcloud}
\alias{simple_wordcloud}
\title{Simple Wordcloud}
\usage{
simple_wordcloud(text)
}
\arguments{
\item{text}{A vector of words}
}
\description{
Wordcloud with no parameters
}
\examples{
simple_wordcloud(lorem_ipsum())

}
